fix(perform): clippy casts, rustfmt, and the HUD toggle's browser gate
- perform_window tests: drop two same-type casts clippy rejects under
  -D warnings; rustfmt reflow of the perform:closed emit.
- The drawer's Mode HUD switch is no longer desktop-gated: the browser
  two-tab flow (/?perform=1) honors it too — verified live, the flash
  ('LED Matrix') rendered on the output tab the moment the mode landed.

+++ src-tauri/src/perform_window.rs
@@ -300,8 +300,8 @@ mod tests {
         let m = mon(1, 1920, 1080, false);
         let (x, y, w, h) = windowed_rect(&m);
         assert_eq!((w, h), (1152, 648));
-        assert_eq!(x, 2560 + (1920 - 1152) as i32 / 2);
-        assert_eq!(y, (1080 - 648) as i32 / 2);
+        assert_eq!(x, 2560 + (1920 - 1152) / 2);
+        assert_eq!(y, (1080 - 648) / 2);
     }
 
     #[test]
